More casts

+++ flamedisx/core.py
@@ -394,8 +394,10 @@ class ERSource:
         # these four are (n_events, |nq|) tensors
         _nq_1d = self.domain('nq')
         rate_nq = self.rate_nq(_nq_1d)
-        pel = self.gimme('p_electron', _nq_1d)
-        pel_fluct = self.gimme('p_electron_fluctuation', _nq_1d)
+        pel = self.gimme('p_electron',
+                         tf.cast(_nq_1d, dtype=tf.float64))
+        pel_fluct = self.gimme('p_electron_fluctuation',
+                               tf.cast(_nq_1d, dtype=tf.float64))
 
         # Create tensors with the dimensions of our final result
         # i.e. (n_events, |photons_produced|, |electrons_produced|),
